requirements_core.txt
# Testing and Development Tools
pytest>=7.3.1
pytest-asyncio>=0.21.1
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1
pytest-cov>=4.1.0

# Core System Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.5
pydantic>=2.3.0
blessed>=1.20.0
pandas>=2.0.0
SpeechRecognition>=3.10.0
gTTS>=2.5.0
google-cloud-speech>=2.21.0
moviepy==2.1.2  # Specific version for TTV pipeline compatibility

# Google APIs and Authentication
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
httplib2>=0.22.0

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
keyboard>=0.13.5
pyaudio>=0.2.13
pydub>=0.25.1

# Image and Media Handling
pillow>=10.0.0
python-magic>=0.4.27

# Development dependencies
pylint>=3.0.3  # For code linting and style checking
